# apps/web/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SystemOneStudio Control Plane",
    "description": "Control plane for the zero-code decision model training platform. The prediction surface is served separately per deployment at /v1/predict.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/models/{model_id}/plan": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "训练方案与耗时估算",
        "description": "Computed hyperparameters plus a duration estimate, clearly labelled.\n\n需求方案.txt 5.6 says the advanced values are recommended rather than typed,\nso they have to be *computed* rather than hardcoded in the UI. Every figure\nhere is an estimate; ``measured`` is false and ``basis`` says why.",
        "operationId": "training_plan_api_models__model_id__plan_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "rows",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10000,
              "title": "Rows"
            }
          },
          {
            "name": "qlora",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Qlora"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Training Plan Api Models  Model Id  Plan Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/adapter": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "适配器可用性",
        "description": "Whether the adapter for a model is registered.\n\nNo adapter is registered yet: the real ones need torch/PEFT and a GPU. This\nreports that honestly instead of returning a placeholder.",
        "operationId": "adapter_status_api_models__model_id__adapter_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Adapter Status Api Models  Model Id  Adapter Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/model-versions": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "注册模型版本",
        "description": "Register a trained model version and its artifacts.\n\nLineage codes are validated by the ``Lineage`` model, so a malformed\n``sc_``/``ds_``/``mv_`` code is a 422 rather than a broken artifact later.",
        "operationId": "register_model_version_api_model_versions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_register_model_version_api_model_versions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "评测报告",
        "description": "Compute the report, performance block first (需求方案.txt 9.1/9.2).\n\nThe two blocks are returned separately and never merged: the priority order\nis an architectural input, not a reporting preference.",
        "operationId": "evaluate_api_evaluations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/deployments": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "一键部署",
        "description": "Produce the deployment plan and call example.\n\nReturns ``status=pending`` with the exact launch command rather than claiming\na service is running. Starting llama.cpp needs a GPU node and a llama.cpp\ncheckout; neither exists here, and reporting a running service from this\nprocess would be false (AGENTS.md).",
        "operationId": "deploy_api_deployments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeployResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jev/level-claim": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "校验 JEV 兼容等级声明",
        "description": "Validate a compatibility claim before it is written.\n\nL3 requires a comparison test; L2 requires the JEV benchmark, which is not in\nthis repository. Both are refused here rather than noted in a comment.",
        "operationId": "claim_jev_level_api_jev_level_claim_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JevLevelClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Claim Jev Level Api Jev Level Claim Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_register_model_version_api_model_versions_post": {
        "properties": {
          "lineage": {
            "$ref": "#/components/schemas/Lineage"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactRecord"
            },
            "type": "array",
            "title": "Artifacts"
          }
        },
        "type": "object",
        "required": [
          "lineage",
          "artifacts"
        ],
        "title": "Body_register_model_version_api_model_versions_post"
      },
...
      "DeployRequest": {
        "properties": {
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "artifact_path": {
            "type": "string",
            "title": "Artifact Path"
          },
          "quant_level": {
            "$ref": "#/components/schemas/QuantLevel",
            "default": "Q4_K_M"
          },
          "port": {
            "type": "integer",
            "title": "Port",
            "default": 8080
          },
          "concurrency": {
            "type": "integer",
            "title": "Concurrency",
            "default": 8
          },
          "gpu_layers": {
            "type": "integer",
            "title": "Gpu Layers",
            "default": 0
          },
          "jev_format_compat": {
            "type": "boolean",
            "title": "Jev Format Compat",
            "default": true
          },
          "jev_training_compat": {
            "type": "boolean",
            "title": "Jev Training Compat",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "model_version",
          "artifact_path"
        ],
        "title": "DeployRequest"
      },
      "DeployResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/RunState"
          },
          "api_url": {
            "type": "string",
            "title": "Api Url"
          },
          "docs_url": {
            "type": "string",
            "title": "Docs Url"
          },
          "curl": {
            "type": "string",
            "title": "Curl"
          },
          "launch_plan": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Launch Plan"
          },
          "jev_format_compat": {
            "type": "boolean",
            "title": "Jev Format Compat"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "status",
          "api_url",
          "docs_url",
          "curl",
          "launch_plan",
          "jev_format_compat",
          "notes"
        ],
        "title": "DeployResponse"
      },
      "EvaluationRequest": {
        "properties": {
          "y_true": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "title": "Y True"
          },
          "y_pred": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "title": "Y Pred"
          },
          "scores": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scores"
          },
          "format_compliant": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Format Compliant"
          },
          "format_total": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Format Total"
          },
          "contributions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Contributions"
          },
          "latency_samples": {
            "items": {
              "$ref": "#/components/schemas/LatencySample"
            },
            "type": "array",
            "title": "Latency Samples"
          },
          "quant_level": {
            "$ref": "#/components/schemas/QuantLevel",
            "default": "Q4_K_M"
          },
          "concurrency": {
            "type": "integer",
            "title": "Concurrency",
            "default": 1
          },
          "wall_clock_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wall Clock S"
          },
          "peak_memory_mb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Memory Mb"
          }
        },
        "type": "object",
        "required": [
          "y_true",
          "y_pred"
        ],
        "title": "EvaluationRequest",
        "description": "Evaluation input.\n\n``format_compliant`` / ``format_total`` are explicit because 格式合规率 is a\ntracked product indicator (需求方案.txt 14.2). They cannot be inferred from\nthe request: a caller supplying predictions says nothing about whether the\nmodel's raw output parsed. Omit them and the metric reports 未测量."
      },
      "EvaluationResponse": {
        "properties": {
          "performance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Performance"
          },
          "effect": {
            "additionalProperties": true,
            "type": "object",
            "title": "Effect"
          },
          "top_factors": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Top Factors"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "performance",
          "effect",
          "top_factors",
          "notes"
        ],
        "title": "EvaluationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JevLevelClaimRequest": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "level": {
            "$ref": "#/components/schemas/JevCompatLevel"
          },
          "evidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "level"
        ],
        "title": "JevLevelClaimRequest"
      },
...
      "Lineage": {
        "properties": {
          "scene": {
            "type": "string",
            "title": "Scene",
            "description": "Scene code, e.g. sc_v3"
          },
          "dataset": {
            "type": "string",
            "title": "Dataset",
            "description": "Dataset code, e.g. ds_v2"
          },
          "synth": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synth",
            "description": "Synthesis run code, e.g. syn_v5. None when synthesis was skipped."
          },
          "base_model": {
            "type": "string",
            "title": "Base Model",
            "description": "Base model id, e.g. qwen2.5-3b-instruct"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "description": "Training method id, e.g. jev_lora_dpo"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "description": "Model version code, e.g. mv_0017"
          }
        },
        "type": "object",
        "required": [
          "scene",
          "dataset",
          "base_model",
          "method",
          "model_version"
        ],
        "title": "Lineage",
        "description": "Immutable provenance snapshot attached to a model version.\n\nCodes follow the naming style in 需求方案.txt section 5.7.1, e.g.\n``sc_v3`` / ``ds_v2`` / ``syn_v5`` / ``mv_0017``."
      },
...
      "ModelVersionResponse": {
        "properties": {
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "lineage": {
            "$ref": "#/components/schemas/Lineage"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactRecord"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "model_version",
          "lineage",
          "artifacts",
          "created_at",
          "notes"
        ],
        "title": "ModelVersionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}